- Add eslintrc - add editorconfig - update .gitignore - add .nvmrc to make sure folks and CI tools (like netlify) are on consistent versions of node - add prettierignore and prettierrc for code style - add config.js for basic central config - npm install esm so we can make use of es6 imports in node - add scripts to package.json. Note, to use es6 imports we need these scripts instead of `gatsby develop` - import the fragment into createPosts - moved Post template into a directory with an index.js and a data.js file. The data.js defines the graphql fragment for the template.
