+++ packages/twentynineteen-gatsby-theme/.nvmrc
@@ -1,0 +1,1 @@
+11.10.1
