+++ packages/twentynineteen-gatsby-theme/config.js
@@ -1,0 +1,5 @@
+const config = {
+    wordPressUrl: `https://demo.wpgraphql.com`,
+};
+
+export default config
