+++ packages/twentynineteen-gatsby-theme/src/templates/Post/data.js
@@ -1,0 +1,8 @@
+export const PostContentFragment = `
+    fragment PostContentFragment on Post { 
+      id
+      postId
+      title
+      content
+    }
+`;
